backend/pyproject.toml
[project]
name = "legal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MawiyaManzar", email = "mawiyamanzar2028@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi>=0.121.2",
    "langchain>=1.0.7",
    "langchain-chroma>=1.0.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.3",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
legal-agent = "legal_agent:main"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
